requirements.txt
flask>=3.0,<4.0
flask-cors>=5.0,<7.0
pandas>=2.1,<3.0
numpy>=1.26,<2.0
scipy>=1.12,<2.0
scikit-learn>=1.4,<2.0

# V2 dependencies (Leiden clustering for hybrid approach)
python-igraph>=0.11,<1.0
leidenalg>=0.10,<1.0
google-auth>=2.0,<3.0
requests>=2.31,<3.0
protobuf
# GCP persistence (Cloud Run + GCS + Firestore)
google-cloud-storage>=2.16,<3.0
google-cloud-firestore>=2.16,<3.0